requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.28.62
botocore==1.31.62
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2